workers/audio/requirements.txt
redis>=5.0.0
boto3>=1.34.0       # Cloudflare R2 via S3-compatible API
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
librosa>=0.10.2
soundfile>=0.12.1
audioread>=3.0.1
pyrubberband>=0.3.0
numpy>=1.26.0
pytest>=8.0.0
